No commit message

+++ grab.sh
@@ -1,9 +1,9 @@
-base=tb_template
+base=tb_i2c
 source=nexys_a7.runs/impl_1/top_level_wrapper
 dest=bitstream
 
 mkdir $dest 2>/dev/null
 
 cp ${source}.bit ${dest}/${base}.bit
-cp ${source}.ltx ${dest}/${base}.ltx
+test -f ${source}.ltx && cp ${source}.ltx ${dest}/${base}.ltx
